renaming all screens folders
