+++ src/features/superAdminUser/userList.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import {
   Table,
   TableBody,
